fix(python): scope checkout caches to the specific project
Since the template uses the `mozilla` trust domain, it's expected for
workers to run tasks for multiple unrelated projects. Therefore we don't
want to share a checkout cache between them (and especially not call
them all `src`).

This scopes the checkout cache to just the specific project as well as
renames it from `src` -> `${project}`.